backend/requirements.txt
django
djangorestframework
django-cors-headers
psycopg[binary]
gunicorn[standard]
whitenoise
djangorestframework-simplejwt